agregamos la funcionalidad de obtener porteros modificarlos y eliminarlos

+++ src/main/java/com/grupo6/persistence/repository/PorteroRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.grupo6.persistence.model.Portero;
 
 @Repository
-public interface PorteroRepository extends BaseRepository <Portero, Integer>{
+public interface PorteroRepository extends BaseRepository <Portero, Long>{
 
 	Optional<Portero> findByCedula(String cedula);
